Aligned SCIM tests and model to last updates to ListResponseDTO

+++ iam-login-service/src/main/java/it/infn/mw/iam/api/tokens/service/DefaultAccessTokenService.java
@@ -72,7 +72,7 @@ public class DefaultAccessTokenService implements TokenService<AccessToken> {
     return tokenRepository.findAllValidAccessTokens(new Date(), op);
   }
 
-  private int countAllValidTokens() {
+  private long countAllValidTokens() {
 
     return tokenRepository.countValidAccessTokens(new Date());
   }
@@ -82,7 +82,7 @@ public class DefaultAccessTokenService implements TokenService<AccessToken> {
     return tokenRepository.findValidAccessTokensForUser(userId, new Date(), op);
   }
 
-  private int countAllValidTokensForUser(String userId) {
+  private long countAllValidTokensForUser(String userId) {
 
     return tokenRepository.countValidAccessTokensForUser(userId, new Date());
   }
@@ -93,7 +93,7 @@ public class DefaultAccessTokenService implements TokenService<AccessToken> {
     return tokenRepository.findValidAccessTokensForClient(clientId, new Date(), op);
   }
 
-  private int countAllValidTokensForClient(String clientId) {
+  private long countAllValidTokensForClient(String clientId) {
 
     return tokenRepository.countValidAccessTokensForClient(clientId, new Date());
   }
@@ -104,12 +104,12 @@ public class DefaultAccessTokenService implements TokenService<AccessToken> {
     return tokenRepository.findValidAccessTokensForUserAndClient(userId, clientId, new Date(), op);
   }
 
-  private int countAllValidTokensForUserAndClient(String userId, String clientId) {
+  private long countAllValidTokensForUserAndClient(String userId, String clientId) {
 
     return tokenRepository.countValidAccessTokensForUserAndClient(userId, clientId, new Date());
   }
 
-  private ListResponseDTO<AccessToken> buildCountResponse(int countResponse) {
+  private ListResponseDTO<AccessToken> buildCountResponse(long countResponse) {
 
     return new ListResponseDTO.Builder<AccessToken>().totalResults(countResponse)
         .resources(Collections.emptyList()).startIndex(1).itemsPerPage(0).build();
@@ -148,7 +148,7 @@ public class DefaultAccessTokenService implements TokenService<AccessToken> {
 
     if (isCountRequest(pageRequest)) {
 
-      int count = countAllValidTokens();
+      long count = countAllValidTokens();
       return buildCountResponse(count);
     }
 
@@ -163,7 +163,7 @@ public class DefaultAccessTokenService implements TokenService<AccessToken> {
 
     if (isCountRequest(pageRequest)) {
 
-      int count = countAllValidTokensForUser(userId);
+      long count = countAllValidTokensForUser(userId);
       return buildCountResponse(count);
     }
 
@@ -178,7 +178,7 @@ public class DefaultAccessTokenService implements TokenService<AccessToken> {
 
     if (isCountRequest(pageRequest)) {
 
-      int count = countAllValidTokensForClient(clientId);
+      long count = countAllValidTokensForClient(clientId);
       return buildCountResponse(count);
     }
 
@@ -193,7 +193,7 @@ public class DefaultAccessTokenService implements TokenService<AccessToken> {
 
     if (isCountRequest(pageRequest)) {
 
-      int count = countAllValidTokensForUserAndClient(userId, clientId);
+      long count = countAllValidTokensForUserAndClient(userId, clientId);
       return buildCountResponse(count);
     }
